* Renamed some files * added some enums * Set StructLayout Size properties * corrected some sizes * Removed obsolete HasToStringOverride attributes

+++ Models/Enums/Act.cs
@@ -1,12 +1,10 @@
-using System;
-
 namespace SoE.Models.Enums
 {
 	/// <summary>
 	/// The acts of the game
 	/// </summary>
 	/// <remarks>2 bytes</remarks>
-	public enum Act: UInt16
+	public enum Act: byte
 	{
 		Prehistoria = 0, 
 		Antiqua = 2,
