+++ Models/Enums/FlyingMachineType.cs
@@ -1,12 +1,10 @@
-using System;
-
 namespace SoE.Models.Enums
 {
 	/// <summary>
 	/// The type of flying machine
 	/// </summary>
 	/// <remarks>1 byte</remarks>
-	public enum FlyingMachineType : UInt16
+	public enum FlyingMachineType : byte
 	{
 		None = 0, 
 		Windwalker = 1,
